+++ aa_forum/__init__.py
@@ -2,5 +2,5 @@
 A couple of variables to use throughout the app
 """
 
-__version__ = "1.10.0"
+__version__ = "1.11.0"
 __title__ = "Forum"