+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [In Development] - Unreleased
 
 
+## [1.11.0] - 2022-08-24
+
+### Added
+
+- Default forum settings to the admin backend
+- `related_name` to foreign keys that where missing them
+
+### Changed
+
+- `Setting` model is now a singleton to prevent multiple settings of the same type
+
+
 ## [1.10.0] - 2022-08-03
 
 ### Fixed
@@ -22,14 +34,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- "New Category" form in admin page visually enhanced
+- "New Category" form in the admin page visually enhanced
 - Contribution guidelines updated
 - CSS modernized
 - App CSS and JS moved to bundled HTML templates
 - Minimum requirements:
   - Python >= 3.8
   - Alliance Auth >= 2.15.1
-  - Django ckEditor >= 6.4.2
+  - Django CKEditor >= 6.4.2
 
 ### Removed